Fix SonarQube Cloud issues across multiple rules

+++ app/src/main/java/com/simplecity/amp_library/ui/views/SnowfallView.java
@@ -167,13 +167,13 @@ public class SnowfallView extends View {
         for (int i = 0; i < numFlakes; i++) {
             final double angle = toRadians(lerp(MIN_ANGLE, MAX_ANGLE, snowRng.nextDouble()));
             final float speed = lerp(MIN_SPEED, MAX_SPEED, snowRng.nextFloat());
-            final float velX = (float) ((double) speed * cos(angle));
-            final float velY = (float) ((double) speed * sin(angle));
+            final float velX = (float) (speed * cos(angle));
+            final float velY = (float) (speed * sin(angle));
             final float size = lerp(MIN_SIZE, MAX_SIZE, snowRng.nextFloat());
-            final float startX = lerp(0f, (float) getWidth(), snowRng.nextFloat());
-            float startY = lerp(0f, (float) getHeight(), snowRng.nextFloat());
-            startY -= (float) getHeight() - size;
-            final int alpha = (int) lerp((float) MIN_ALPHA, (float) MAX_ALPHA, snowRng.nextFloat());
+            final float startX = lerp(0f, getWidth(), snowRng.nextFloat());
+            float startY = lerp(0f, getHeight(), snowRng.nextFloat());
+            startY -= getHeight() - size;
+            final int alpha = (int) lerp(MIN_ALPHA, MAX_ALPHA, snowRng.nextFloat());
             snowflakes.add(new Snowflake(startX, startY, velX, velY, size, alpha));
         }
         invalidate();
